+++ plugins/catalog/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-catalog
 
+## 0.6.3
+
+### Patch Changes
+
+- 30c2fdad2: Exports `CatalogLayout` and `CreateComponentButton` for catalog customization.
+- e2d68f1ce: Truncate long entity names on the system diagram
+- d2d42a7fa: Fix for Diagram component using hard coded namespace.
+- 2ebc430c4: Export `CatalogTableRow` type
+- Updated dependencies
+  - @backstage/plugin-catalog-react@0.2.3
+  - @backstage/catalog-model@0.8.3
+  - @backstage/core@0.7.13
+
 ## 0.6.2
 
 ### Patch Changes